Creda_Fastapi/pyproject.toml
[project]
name = "creda-backend"
version = "2.0.0"
description = "CREDA — AI-powered multilingual financial advisory backend for Indian users"
requires-python = ">=3.11,<3.14"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "CREDA Team" }]

dependencies = [
    # ── Framework ──
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "jinja2>=3.1",
    "aiofiles>=23.2",

    # ── Database & persistence ──
    "sqlmodel>=0.0.16",
    "aiosqlite>=0.20",

    # ── LangGraph agent orchestration ──
    "langgraph>=0.2",
    "langchain-core>=0.2",
    "langchain-groq>=0.1",

    # ── LLM ──
    "groq>=0.4",

    # ── Financial math ──
    "casparser>=0.3",
    "pyxirr>=0.10",
    "yfinance>=0.2",
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",

    # ── RAG ──
    "chromadb>=0.5",
    "sentence-transformers>=2.6",

    # ── PDF ──
    "pdfplumber>=0.10",
    "pymupdf>=1.23",

    # ── Utilities ──
    "python-dateutil>=2.8",
    "requests>=2.31",
]

[project.optional-dependencies]
multilingual = [
    # Only needed when running the Multilingual Service (Port 8000)
    "torch>=2.1",
    "torchaudio>=2.1",
    "transformers>=4.36",
    "soundfile>=0.12",
    "parler-tts @ git+https://github.com/huggingface/parler-tts.git",
    "IndicTransToolkit",
    "sentencepiece>=0.2",
    "sacremoses>=0.1",
    "gtts>=2.5",
    "pydub>=0.25",
]

dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
creda-gateway = "app:main"
creda-finance = "fastapi2_finance:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents"]

# ── Ruff (linter + formatter) ──
[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
quote-style = "double"

# ── Pytest ──
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
